moved history_to_csv and history_list_to_csv to util.py
